Update 11 of Word Doc
* Updates the book to Update 4

Update 4 also sports downloadable dashboards + views.

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Moved vROPs content into ZIP file

If a user clicks any of the content files, the browser tries to open them in a new tab.

As a ZIP the browser will treat it like a blob for download.

* Update 5 of the book

This update sports the PowerPoint version

* Update 5 of the book

This update sports the PowerPoint version

* Updated the content of the page

Wlll be supplying the actual download file after this.

* Update 6 of the book

Massive update on the book, preparing for printed version, supporting vR Ops 8.6, and laying the groundwork for esxtop

* Update 7 (Nov 2021)

Update 7 of the book supersedes the old version.

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Update 8 (Dec 2021)

Preparing for the esxtop release when PJ is ready in Jan 2022 and Horizon update

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Update 9 (Jan 2022) of the book

Just change 1 MS word file

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Update 10 (Feb 2022)

A lot of miscellaneous updates (add, delete, fix). Total page exceeded 700 for the first time.

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Update 11! March 2022

Update 11 brought it a lot of minor (and some not so minor changes throughout). I could not update the MS Word version number as many folks are using it.

Co-Authored-By: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>

* Update 11 of the MS Word doc

Various updates throughout the book

Co-authored-by: Stellios Williams <24645566+TheNewStellW@users.noreply.github.com>